Move some PostgreSQL-specific behaviors out of Redshift base class
